requirements.txt
Django==3.2.9
django-cors-headers==3.8.0
django-crispy-forms==1.13.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0